+++ src/models/Settings.php
@@ -39,7 +39,7 @@ class Settings extends Model
     /**
      * @inheritdoc
      */
-    public function init()
+    public function init(): void
     {
         parent::init();